+++ src/picasapy/app/i18n/picasapy_hu.ts
@@ -3866,6 +3866,16 @@ Egy gyűjtemény megnyitásához kattintson duplán a nevére, vagy kattintson a
         <source>Kiadások megtekintése a GitHubon</source>
         <translation>Kiadások megtekintése a GitHubon</translation>
     </message>
+    <message>
+        <location filename="../qml/PicasaPy/MainToolbar.qml" line="0"/>
+        <source>New Album</source>
+        <translation>Új album</translation>
+    </message>
+    <message>
+        <location filename="../qml/PicasaPy/MainToolbar.qml" line="0"/>
+        <source>Create a new album</source>
+        <translation>Új album létrehozása</translation>
+    </message>
 </context>
 <context>
     <name>PerfMonitorPanel</name>